requirements.txt
mysql-connector-python==9.1.0
pandas==2.2.3
openpyxl==3.1.5
sqlalchemy==2.0.36
bcrypt==4.2.0